+++ projects/ngx-stripe/src/lib/components/payment-element.component.ts
@@ -79,7 +79,7 @@ export class StripePaymentElementComponent implements OnInit, OnChanges, OnDestr
         ...(this.elementsOptions || {}),
         ...(this.appearance ? { appearance: this.appearance } : {}),
         ...(this.clientSecret ? { clientSecret: this.clientSecret } : {})
-      }).toPromise();
+      } as StripeElementsOptions).toPromise();
       updateElements = true;
     }
 
@@ -111,7 +111,7 @@ export class StripePaymentElementComponent implements OnInit, OnChanges, OnDestr
         ...(this.elementsOptions || {}),
         ...(this.appearance ? { appearance: this.appearance } : {}),
         ...(this.clientSecret ? { clientSecret: this.clientSecret } : {})
-      }).toPromise();
+      } as StripeElementsOptions).toPromise();
       this.createElement(options);
 
       this.state = 'ready';